working.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()
GROQ_API_KEY = os.getenv("groq_api_key")

HEADERS = {
    "Authorization": f"Bearer {GROQ_API_KEY}",
    "Content-Type": "application/json"
}

def generate_question(role="software developer"):
    prompt = f"Generate a behavioral interview question for a {role}."
    data = {
        "model": "llama3-8b-8192",
        "messages": [
            {"role": "user", "content": prompt}
        ]
    }
    response = requests.post(
        "https://api.groq.com/openai/v1/chat/completions",
        headers=HEADERS,
        json=data
    )
    return response.json()["choices"][0]["message"]["content"].strip()

def analyze_answer(answer):
    promp
