requirements.txt
streamlit
requests
python-dotenv
SpeechRecognition
gTTS
playsound==1.2.2
